requirements.txt
behave
selenium
pylint
flake8
dotenv
requests
beautifulsoup4
pytest
allure-pytest
loguru
pytest-xdist
coverage
Pillow
axe-selenium-python
allure-behave
numpy
opencv-python
pandas
scikit-learn
openai
